# Avoid allocating a string for each struct field when unmarshalling