This reuses the allocated memory for the struct field by switchting from `BulkString` to `BulkStringBytes` and using the `map[string([]byte)]` optimization introduced in Go 1.3.

```
name                       old time/op    new time/op    delta
AnyUnmarshalRESP/Struct-8    1.68µs ± 1%    1.35µs ± 1%  -19.73%  (p=0.000 n=9+8)

name                       old alloc/op   new alloc/op   delta
AnyUnmarshalRESP/Struct-8      112B ± 0%       96B ± 0%  -14.29%  (p=0.000 n=10+10)

name                       old allocs/op  new allocs/op  delta
AnyUnmarshalRESP/Struct-8      7.00 ± 0%      4.00 ± 0%  -42.86%  (p=0.000 n=10+10)
```